// src/types/accessorize.tsx
interface Description {
    title: string;
    text: string[];
  }

export interface Accessorize {
    id: string;
    category: string;
    namespaceId: string;
    name: string;
    capacityAvailable: string[];
    capacity: string;
    priceRegular: number;
    priceDiscount: number;
    colorsAvailable: string[];
    color: string;
    images: string[];
    description: Description[];
    screen: string;
    resolution: string;
    processor: string;
    ram: string;
    cell: string[];
  }
  
  